refactor cross chain initialization (#2194)
