+++ protocol/governance/test/integration/Evaluation.test.ts
@@ -11,13 +11,6 @@ describe('Evaluation', function () {
     await fastForwardTo(schedule.endDate.add(10).toNumber(), provider);
   };
 
-  before('register emitters', async function () {
-    await chains.mothership.GovernanceProxy.connect(chains.mothership.signer).setRegisteredEmitters(
-      [10002],
-      [chains.mothership.GovernanceProxy.address]
-    );
-  });
-
   describe('no nominees', () => {
     it('should jump to nomination period', async () => {
       const { mothership } = chains;
